+++ avr/cores/megacommand/MCL/MDTempoTrack.h
@@ -29,7 +29,7 @@ public:
   void load_immediate(uint8_t tracknumber, SeqTrack *seq_track);
 
   virtual uint16_t get_track_size() { return sizeof(MDTempoTrack); }
-  virtual uint32_t get_region() { return BANK1_MDTEMPO_TRACK_START; }
+  virtual uint16_t get_region() { return BANK1_MDTEMPO_TRACK_START; }
 
   virtual uint8_t get_model() { return MDTEMPO_TRACK_TYPE; }
   virtual uint8_t get_device_type() { return MDTEMPO_TRACK_TYPE; }
